+++ Assets/PotatoUtil/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.0] - 2021-01-14
+### Added
+- `NonDrawingGraphic` and `NonDrawingGraphicEditor` courtesy of Slipp Douglas Thompson's [Gist](https://gist.github.com/capnslipp/349c18283f2fea316369).
+- `PotatoUtil.Editor` namespace for Editor only scripts and utilities.
+
 ## [1.4.1] - 2021-01-02
 ### Fixed
 - `Base32` defect where `Parse` would always fail with correct length strings